Rewrite OPEND using blueprint
