+++ .gitignore
@@ -1,0 +1,17 @@
+# Windows:
+Thumbs.db
+ehthumbs.db
+Desktop.ini
+
+# Python
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Bower dependency directory (https://bower.io/)
+#static/bower_components/
+static/node_modules/
+
+
+
