--- app/templates/base.html
@@ -1,21 +1,0 @@
-{% extends "bootstrap/base.html" %}
-<html>
-  <head>
-    <title>Welcome to OPEND</title>
-  </head>
-  <body>
-    <hr>
-    {% block content %}{% endblock %}
-  </body>
-</html>
-
-<script>
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-  })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-
-  ga('create', 'UA-71696336-4', 'auto');
-  ga('send', 'pageview');
-
-</script>
