--- app/templates/mainform.html
@@ -1,69 +1,0 @@
-<script>
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-  })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-
-  ga('create', 'UA-71696336-4', 'auto');
-  ga('send', 'pageview');
-</script>
-
-<style>
-	body {background-color:#D5D8DC}
-</style>
-<h1 style="color:#3d475c;font-size:75px;font-weight:bold;text-align:left;margin-left:10px">OPEND</h1>	  
-<div style="color:#3d475c;font-size:15px;font-weight:normal;text-align:left;margin-left:14px">A volunteer project of <a href="https://www.code.sydney" target="_blank">Code.Sydney</a></div>	  
-
-      <div style="font-size:20px;text-align:left;margin-left:10px;color:red">
-      {% for message in form.InputSuburb.errors %}
-         <div>{{ message }}</div>
-      {% endfor %}
-	    </div> 	  
-	  
-      <form action = "" method="POST" name="mainform">
-           <fieldset>
-            <br>
-            {{ form.hidden_tag() }}
-            
-            <div style = "font-size:30px;font-weight:normal;text-align:left;margin-left:10px">
-			   <h2 style="color:#3d475c;font-size:40px;font-weight:bold;text-align:left">Suburb Information</h2>			  			   
-        	   
-   			 {{ form.InputSuburb.label }}<br>
-         <!-- {{ form.InputAddress }} -->
-   
-<head>
-<link href="//code.jquery.com/ui/1.10.2/themes/smoothness/jquery-ui.css" rel="Stylesheet"></link>
-<script src="//code.jquery.com/jquery-2.2.0.min.js"></script>
-<script src="//code.jquery.com/ui/1.10.2/jquery-ui.js" ></script>            
-<script type="text/javascript">
-$(function() {
-    $("#autocomplete").autocomplete({
-        source:function(request, response) {
-            $.getJSON("{{url_for('autocomplete')}}",{
-                q: request.term, // in flask, "q" will be the argument to look for using request.args
-            }, function(data) {
-                response(data.matching_results); // matching_results from jsonify
-            });
-        },
-        minLength: 1,
-        select: function(event, ui) {
-            console.log(ui.item.value); // not in your question, but might help later
-        }
-    });
-})
-</script>
-</head>
-
-<body>
-    <div>
-        <input name="InputSuburb" type="text" id="autocomplete" class="form-control input-lg"/>
-    </div>
-</body>		   			   			  
-			  <br>
-			  {{ form.Submit1 }}
-            </div> 
-      </form>
-	  
-<br>
-<br>
-<div style="color:#3d475c;font-size:15px;font-weight:normal;text-align:left;margin-left:14px">This work is licensed under a <a href="https://creativecommons.org/licenses/by/4.0/" target="_blank">Creative Commons Attribution License</a></div>
